fix(o2_consumer_health): recognize non-first o2 tanks
The purpose of this change is to correct a defect that caused the
o2/health system to short-circut the entire check for o2 tanks when any
item in a container was not recognized as an o2 tank. This meant as soon
as a battery was "in front" of an o2 tank, that o2 tank would never get
used.